作了2个优化，一是根据原图1360*800的尺寸进行框框的截取，而非粗暴的缩放；而是调整训练/测试/验收的随机，应该以box框框列表进行随机，而非id

+++ train.py
@@ -14,6 +14,7 @@
 import os
 import re
 import time
+import math
 import shutil
 import numpy as np
 
@@ -21,10 +22,6 @@ import numpy as np
 #import tensorflow as tf
 import tensorflow.compat.v1 as tf
 tf.disable_v2_behavior()
-
-tf_config = tf.ConfigProto()  
-tf_config.gpu_options.allow_growth = True  
-session = tf.Session(config=tf_config) 
 
 import core.utils as utils
 from tqdm import tqdm
@@ -51,7 +48,13 @@ class YoloTrain(object):
         self.trainset            = Dataset('train')
         self.testset             = Dataset('test')
         self.steps_per_period    = len(self.trainset)
-        self.sess                = tf.Session(config=tf.ConfigProto(allow_soft_placement=True))
+
+        tf_config = tf.ConfigProto(allow_soft_placement=True)
+        
+        #lws
+        tf_config.gpu_options.allow_growth = True  
+
+        self.sess                = tf.Session(config=tf_config)
 
         with tf.name_scope('define_input'):
             self.input_data   = tf.placeholder(dtype=tf.float32, name='input_data')
@@ -144,19 +147,21 @@ class YoloTrain(object):
             print('=> Now it starts to train YOLOV3 from scratch ...')
             from_scratch = True
             self.first_stage_epochs = 0
-
+        
+        # checkpont列表
+        ckpt_list = []
+        
         last_epoch = 0
         if not from_scratch:
             try:
-                found = re.sub(r'.*yolov3_test_loss_.*.ckpt-(\d+)', r'\1', self.initial_weight)
-                last_epoch = int(found)
+                found = re.sub(r'.*yolov3_test_loss_(.*)\.ckpt-(\d+)', r'\1:\2', self.initial_weight)
+                foundList = found.split(':')
+                last_test_loss, last_epoch = float(foundList[0]), int(foundList[1])
+                ckpt_list.append((self.initial_weight+" @", last_test_loss))
             except:
-                print("found:", found)
                 print("=> faield to capture the last epoch in %s" % self.initial_weight)
                 last_epoch = 0
 
-        # checkpont列表
-        ckpt_list = []
         
         # 总的epoch
         total_epoch = self.first_stage_epochs+self.second_stage_epochs
@@ -201,26 +206,34 @@ class YoloTrain(object):
 
                 test_epoch_loss.append(test_step_loss)
 
-            train_epoch_loss, test_epoch_loss = np.mean(train_epoch_loss), np.mean(test_epoch_loss)
-            ckpt_file = "./checkpoint/yolov3_test_loss_%.4f.ckpt" % test_epoch_loss
-            log_time = time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(time.time()))
-            print("=> Epoch: %d(%d/%d) Time: %s Train loss: %.2f Test loss: %.2f Saving %s ..."
-                            %(global_epoch, epoch, total_epoch, log_time, train_epoch_loss, test_epoch_loss, ckpt_file))
-            self.saver.save(self.sess, ckpt_file, global_step=global_epoch)
-            
             # 记录该快照
-            ckpt_list.append((ckpt_file, test_epoch_loss))
-            ckpt_list.sort(key=lambda elem: elem[1])
             limit = 5
-            if len(ckpt_list) > limit:
-                ckpt_list, remove_list = ckpt_list[:limit], ckpt_list[limit:]
-                for elem in remove_list:
-                    fname = elem[0]    
-                    cmd = "rm "+fname+"*"
-                    print("calling ", cmd)
-                    os.system(cmd)
-            if len(ckpt_list) > 0:
-                print("current best:", ckpt_list[0][0])
+            
+            # 如果列表未满，且小于最大的损失率才添加
+            train_epoch_loss, test_epoch_loss = np.mean(train_epoch_loss), np.mean(test_epoch_loss)
+            
+            if test_epoch_loss != None and math.isnan(float(test_epoch_loss)) == False:
+                # 开始存档
+                ckpt_file = "./checkpoint/yolov3_test_loss_%.4f.ckpt" % test_epoch_loss
+                log_start_time = time.localtime(time.time())
+                log_time = time.strftime('%Y-%m-%d %H:%M:%S', log_start_time)
+                print("=> Epoch: %d(%d/%d) Time: %s Train loss: %.2f Test loss: %.2f Saving %s ..."
+                                %(global_epoch, epoch, total_epoch, log_time, train_epoch_loss, test_epoch_loss, ckpt_file))
+                self.saver.save(self.sess, ckpt_file, global_step=global_epoch)
+                 
+                ckpt_list.append((ckpt_file+"-"+str(global_epoch), test_epoch_loss))
+                ckpt_list.sort(key=lambda elem: elem[1])
+                if len(ckpt_list) > limit:
+                    ckpt_list, remove_list = ckpt_list[:limit], ckpt_list[limit:]
+                    for elem in remove_list:
+                        fname = elem[0]    
+                        cmd = "rm "+fname+"*"
+                        print("calling ", cmd)
+                        os.system(cmd)
+                if len(ckpt_list) > 0:
+                    for k in range(len(ckpt_list)-1, -1, -1):
+                        elem = ckpt_list[k]
+                        print("current ranking:", elem[0])
 
 
 if __name__ == '__main__': YoloTrain().train()
